Updated files and renamed sections for bug fix
