+++ App.js
@@ -10,13 +10,15 @@ import {UserComponents} from './src/components/UserComponents';
 import HomeScreen from './src/screens/HomeScreen';
 import SigninScreen from './src/screens/visitorScreens/SigninScreen';
 import SignupScreen from './src/screens/visitorScreens/SignupScreen';
-import PendingPaymentsScreen from './src/screens/userScreens/PendingPaymentsScreen';
-import MakePaymentScreen from './src/screens/userScreens/MakePaymentScreen';
-import PaymentSelectScreen from './src/screens/userScreens/PaymentSelectScreen';
-import PaymentConfirmationScreen from './src/screens/userScreens/PaymentConfirmationScreen';
-import PaidConfirmationScreen from './src/screens/userScreens/PaidConfirmationScreen';
 import UserOptionsScreen from './src/screens/userScreens/UserOptionsScreen';
 import ProfileScreen from './src/screens/userScreens/ProfileScreen';
+
+// Finance Screens
+import BillOverviewsScreen from './src/screens/financeScreens/BillOverviewsScreen';
+import PendingPaymentsScreen from './src/screens/financeScreens/PendingPaymentsScreen';
+import MakePaymentScreen from './src/screens/financeScreens/MakePaymentScreen';
+import PaymentSelectScreen from './src/screens/financeScreens/PaymentSelectScreen';
+import PaidConfirmationScreen from './src/screens/financeScreens/PaidConfirmationScreen';
 
 // FireStore
 import auth from '@react-native-firebase/auth';
@@ -65,6 +67,11 @@ const App = () => {
             options={{headerShown: false}}
           />
           <Stack.Screen
+            name="BillOverview"
+            component={BillOverviewsScreen}
+            options={{headerShown: false}}
+          />
+          <Stack.Screen
             name="PayNow"
             component={PendingPaymentsScreen}
             options={{headerShown: false}}
@@ -72,11 +79,6 @@ const App = () => {
           <Stack.Screen
             name="MakePayment"
             component={MakePaymentScreen}
-            options={{headerShown: false}}
-          />
-          <Stack.Screen
-            name="ConfirmPayment"
-            component={PaymentConfirmationScreen}
             options={{headerShown: false}}
           />
           <Stack.Screen
@@ -95,15 +97,14 @@ const App = () => {
           <Stack.Screen
             name="MakeNewPoll"
             component={MakeNewPollScreen}
-            options={{ headerShown: false }}
+            options={{headerShown: false}}
           />
 
           <Stack.Screen
             name="MakeNewEvent"
             component={MakeNewEventScreen}
-            options={{ headerShown: false }}
+            options={{headerShown: false}}
           />
-
 
           {/* Settings and Profile Management */}
           <Stack.Screen
